Test_RAG/requirements.txt
langchain==0.1.0
langchain-community==0.0.13
chromadb==0.4.22
sentence-transformers==2.2.2
python-dotenv==1.0.0
huggingface-hub==0.19.4 